Renaming coordinates.io -> coordinates.data
