+++ pyemma/coordinates/data/featurizer.py
