+++ pyemma/coordinates/data/frames_from_file.py
