+++ pyemma/coordinates/data/interface.py
