+++ pyemma/coordinates/data/writer.py
