+++ pyemma/coordinates/tests/test_featurereader_and_tica_projection.py
@@ -10,7 +10,7 @@ import numpy as np
 import numpy.random
 import mdtraj
 from pyemma.coordinates.api import feature_reader, tica, _TICA as TICA
-from pyemma.coordinates.io.feature_reader import FeatureReader
+from pyemma.coordinates.data.feature_reader import FeatureReader
 from pyemma.util.log import getLogger
 
 log = getLogger('TestFeatureReaderAndTICA')
