+++ pyemma/coordinates/tests/test_tica.py
@@ -6,7 +6,7 @@ Created on 02.02.2015
 import unittest
 import numpy as np
 
-from pyemma.coordinates.io.data_in_memory import DataInMemory
+from pyemma.coordinates.data.data_in_memory import DataInMemory
 from pyemma.coordinates.transform.tica import TICA
 from pyemma.util.log import getLogger
