+++ pyemma/coordinates/tests/test_uniform_time.py
@@ -6,7 +6,7 @@ Created on 09.04.2015
 import unittest
 
 import numpy as np
-from pyemma.coordinates.io.data_in_memory import DataInMemory
+from pyemma.coordinates.data.data_in_memory import DataInMemory
 from pyemma.coordinates.clustering.uniform_time import UniformTimeClustering
